sched/env: add tg_envc in task_group_s to avoid some loops in code
this commit optimize the env api speed

+++ sched/environ/env_unsetenv.c
@@ -61,7 +61,7 @@ int unsetenv(FAR const char *name)
 {
   FAR struct tcb_s *rtcb = this_task();
   FAR struct task_group_s *group = rtcb->group;
-  int idx;
+  ssize_t idx;
 
   DEBUGASSERT(group);
